renamed files names appropriately with undescores
