+++ public/index.html
@@ -49,6 +49,6 @@
     <script src="vendor/d3-5.12.0/d3.js"></script>
     <script src="vendor/d3-legend-2.25.6/d3-legend.min.js"></script>
     <script src="vendor/d3-tip-0.9.1/d3-tip.min.js"></script>
-    <script src="js/controllers/app.controller.js" type="module"></script>
+    <script src="js/controllers/app_controller.js" type="module"></script>
 </body>
 </html>
