+++ public/js/controllers/auth_controller.js
