+++ public/js/controllers/coding_progress_table_controller.js
