+++ public/js/controllers/data_controller.js
@@ -1,4 +1,4 @@
-import { mediadb } from "./app.controller.js";
+import { mediadb } from "./app_controller.js";
 
 export class DataController {
     static updateData(snapshot, data) {
