+++ public/js/controllers/traffic_graphs_controller.js
